third_party/WebKit/LayoutTests/http/tests/dom/EventListener-incumbent-global-1.html

<!DOCTYPE html>
<!--
This test is imported from
https://github.com/w3c/web-platform-tests/blob/master/dom/events/EventListener-incumbent-global-1.sub.html
-->
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe src="//www1.web-platform.test:8080/dom/resources/EventListener-incumbent-global-subframe-1.html"></iframe>
<script>

var t = async_test("Check the incumbent global EventListeners are called with");

onload = t.step_func(function() {
  onmessage = t.step_func_done(function(e) {
    var d = e.data;
    assert_equals(d.actual, d.expected, d.reason);
  });

  frames[0].postMessage("start", "*");
});

</script>
